Add OpenSRPTable to child register page
